requirements.txt
pypresence
psutil
python-dateutil
ttkthemes